\name{NIMBIOS-package}
\alias{NIMBIOS-package}
\alias{NIMBIOS}
\docType{package}
\title{
Support for the Teaching with R tutorial at NIMBIOS
}
\description{
Collaborative editing and some data for the tutorial
}
\details{
\tabular{ll}{
Package: \tab NIMBIOS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-05\cr
License: \tab Use it as you will\cr
}

Load this package to have access to the collaborate() function for
collaborative editing, and to a few data sets.
}
\author{
Daniel Kaplan

Maintainer: Daniel Kaplan <kaplan@macalester.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\examples{
f <- collaborate("sharedfile.Rmd")
data(NHANES-original)
data(NHANES)
}
